requirements.txt
# Core scraping libraries
praw>=7.7.0
tweepy>=4.12.0
requests>=2.28.0
beautifulsoup4>=4.11.0

# Enhanced features
textblob>=0.17.1
python-dotenv>=0.19.0

# CLI and utilities
click>=8.1.0
rich>=12.0.0
tqdm>=4.64.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Testing
pytest>=7.0.0
pytest-mock>=3.8.0
pytest-cov>=4.0.0

# Development
black>=22.0.0
flake8>=5.0.0
mypy>=0.991 